Move macOS bookmark logic to dedicated security_bookmarks module
- Extract macos_file_handler module from main.rs to src/security_bookmarks.rs
- Move test_crash_logging_methods function to new module
- Update all call sites in main.rs and file_io.rs to use new module path
